Implemented dijkstra algorithm and removed redundant argument in method
As it stands, dijkstra's algorithm is trivial. Because the graph class creates a fully connected graph and uses Euclidean distance to compute the edge weights, the shortest distance between a starting node v and an ending node u will always be a direct edge from v to u.

Removed graph pointer argument from calcPathLength method. The method is part of the graph class, so it is able to use the graph's methods without the need for a graph pointer.