Added multithreaded two opt method and class alias
The threadedTwoOpt method provides the same functionality as the twoOpt method, but divides the work between four threads of execution. Each thread runs an instance of the findBestChange method which finds the best improvement (if there is one) over a given interval. (one fourth of the number of vertices) The result is then saved to one of four spots in an array. When all threads are finished, the array is checked and if one of the spots contains an improvement, the best improvement of the four is used.

added alias for TSP_Algos::TSP_Algo_Nearest_Neighbors in order to shorten line length and improve readability