Shortened indentation length and separated class responsibility
Indentation is now two spaces instead of four.

Vertex class has been split into Vertex and VertexEuclid classes.
The Vertex class stores a vertex id and a list of edges to vertices it is connected to.
The VertexEuclid class inherits from Vertex and stores an x and y coordinate to represent its position on a Euclidean plane.