Fixed random vertex button not working
random vertex button added new vertices, but did not connect vertices to
existing ones. The function in makeGraphComplete in Graph class has been
made public so that it can be called by the main window when new random
vertices are added.

vertex.h accidentally not added to repository when the Vertex class was separated from graph.h. vertex.h has now been added to the repository.