Removed unnecessary const reference for primitives
Many of the functions within this project use "const int &arg" to pass
in values. After doing some research, I discovered that primitives such
as ints should not be passed by const reference as it is actually slower
to do it this way than it is to pass them by value.

Some class definitions in the Graph class have been moved to the outside
and made inline.